Safely load thumbnails on different thread
Most of the time, thumbnails are already loaded in memory with
other attributes. In this case, it is only necessary to decode the
embeded thumbnails which can be done in parallel.

If an entity does not have an embeded thumbnail, this procedure
loads and resizes the original image on a different thread. This
implementation uses a thread pool thread for now.