Pass parameters for query evaluation
In particular, it is now possible to pass in a cancellation token
which means the query can be evaluated multiple times and it
does not have to have its own CancellationTokenSource.

This changes the query interface slightly. It now accepts
IExecutableQuery instead of IEnumerable. It is still possible
to convert IEnumerable to IExecutableQuery by using the
MemoryQuery class.