Vastly improve latency of except and intersect queries
To compute an intersection, we can list all items from both queries and
then statically check whether the found entity matches the other query.

In case of an except query, we don't even have to evaluate the other
query. It is sufficient to just list all items from the first query which
don't match the second query.