Fix: add an option to overwrite textual representation of a query
This is considered a fix since a previous commit
(acbf74a6536242cd16d9e2de480e68eb59b748f5) broke this API.
Even though a textual representation of a query can be generated,
it is not always desired. This change shows the query to the user
with the same formatting.