Refactor: move path pattern functions to a custom class
Path pattern is now represented with a custom immutable class. It
contains some useful functions to create derived patterns. Moreover,
pattern regex compilation is now done lazily. This should save some
CPU cycles and memory since, in most situations, it is not used. This
also means that creation of PathPattern is pretty lightweight.