Feat: show current query together with query history in thumbnail grid
This addresses #11 but it does not quite finishes the task. Finished
control should show buttons to go back, forward in history and a button
to go to the parent folder. It uses anchor for layout so it won't be
correct for high DPI settings. This should be easily fixable by setting
the size programatically.