Fix: EntityManager handles modified entities correctly
Modified entities are kept in main memory until the caller decides
what to do with them. The GetEntity method will use this to load
pending entities from memory. This prevents it from loading an
older version of an entity.

Entities can now also be reverted to their initial state. Returning
entities to the modified list can now be done with a method on
returned proxy object.