Preallocate space for the temporary file used for saving attributes
This is done in an effort to speed up the Store algorithm. It seems
that allocating disk space in small blocks is slowing it down. This
change speeds it up by ~17% in my tests, although it is quite
difficult to benchmark this properly. Nevertheless, this brings it
close to the time necessary to copy same files using the xcopy
and robocopy utilities.