Add query suggestions service
This class tries to parse given query up to the given point (the caret
location). If it encounteres an error at the caret location, it captures
current parser context, caret token and a list of expected tokens. This
can be used for generating suggestions.

Cases where the query prefix before the caret position is a valid
query have to be handeled differently. This method won't return
suggestions for them. In these situations though, only keyword
suggestions should be available which are less useful and could even
be distracting.