Change how the state for suggestions is captured
The new algorithm traverses the ATN graph using tokens before the caret
token and collects all alternatives. The old algorithm parsed the query
using the generated parser. That approach ended up with only 1 path
and it wasn't always the correct path to choose (see the failing test in
the previous version).