Thumbnail grid: navigate with arrow keys
It is now possible to select items with arrow keys. If a shift key is
pressed, this will work as a range select (same as in Shift +
mouse click).

This commit includes substantial refactoring of the thumbnail
grid view. Most of the code which handles selection logic has
been factored out of the view into a new class.