Fix: hide suggestion form whenever its form is deactivated
This solution is more complicated but it should handle all cases.
It saves all controls on the path to the root. Whenever a control is
added/removed, all event handlers are automatically
(un)subscribed. The net result is that suggestions are now hidden
properly even if they are inside of a control in a floating window.