Fix: prevent storage from evicting cache entries during query evaluation
Previously, the DeleteOutdated method could run concurrently with
readers. This had some undesired behaviour such as deleting a file
from the database only to store it back in right after that.

The method can still run in parallel with Load method. However, as
long as there is a non-disposed reader (CreateReader), it won't run.