Fix: PathPattern can deal with .. and . patterns
These patterns are evaluated in PathPattern constructor so that
methods of PathPattern don't have to deal with nonnormalized paths.

This changes behaviour of the .. in query evaluation. It now removes
even the ** pattern. This makes it consistent with *, ? and makes it
easier to reason about since the pattern language is now regular.